requirements.txt
# Google Cloud and Agent Development Kit dependencies
google-cloud-pubsub==2.21.1
google-cloud-secretmanager==2.20.0
google-cloud-bigquery==3.12.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
vertexai==1.38.1

# Agent Development Kit
# Note: ADK is not yet publicly available, using placeholder for core dependencies
google-cloud-functions-framework==3.5.0
functions-framework==3.5.0

# HTTP and API tools
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.3
numpy==1.24.3

# Email and communication
sendgrid==6.10.0
twilio==8.10.0

# JSON and data handling
pydantic==2.5.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1